Try to use the embedded cuda lib first
Favor the cudart lookup, and try to use the version we
bundle into the binary if possible before falling
back to other cudart libraries found on the system.
Adds window support.

+++ gpu/gpu_info_cudart.h
@@ -17,9 +17,9 @@ typedef enum cudartDeviceAttr_enum {
 
 typedef void *cudartDevice_t;  // Opaque is sufficient
 typedef struct cudartMemory_st {
-  unsigned long long total;
-  unsigned long long free;
-  unsigned long long used;
+  size_t total;
+  size_t free;
+  size_t used;
 } cudartMemory_t;
 
 typedef struct cudartDriverVersion {
